looks like cid 332 is caused by use of unsigned long instead of the implicit uint32_t type it's using it as.  change getlong() to return 32-bit type so we hopefully avoid all truncation propagation.
svn:revision:50139
svn:branch:trunk
svn:account:brlcad

+++ src/util/sun-pix.c
@@ -109,11 +109,11 @@ Usage: sun-pix [-b -h -i -P -v -C] [sun.bitmap]\n";
 
 #define NET_LONG_LEN 4 /* # bytes to network long */
 
-unsigned long
+uint32_t
 getlong(char *msgp)
 {
     unsigned char *p = (unsigned char *) msgp;
-    unsigned long u;
+    uint32_t u;
 
     u = *p++; u <<= 8;
     u |= *p++; u <<= 8;
